Add a basic load test spec that actually logs folks in.
Requires dropping a file with a list of email addresses that can login
to the specific course before you can run it. With a 4GB allocation
on the selenium server I was able to get through at least 21 concurrent logins.
NOTE: I suspect they're not really running in parallel though since the times
on the server side (using honeycomb) are much less than on the client side
(the standalone selenium server we use the chrome web driver with to execute
the tests). Try using a Grid Hub with nodes to get more confidencxe we're really
running things in parallel instead of just "as fast as we can go".